Autodetec lib, include and bin path for Qt
Until now the path detected was for the includes, it was assuming
that on Linux the Qt binaries are in the path and that the libraries
are part of the system libraries.
This makes clear that we are only using the include path and this is
the first step to implement Qt rules as a toolchain.
For the moment we print the folders for include, lib, and binary that
are detected and this will later be part of the toolchain.