initial setup of bazel toolchain for qt (#28)
The goal is to separate rule implementations from the platform-specific
binaries and settings needed. To start, the toolchain provides
hard-coded paths to the moc, rcc, and uic tools. tools/BUILD provides
separate values for Windows and Linux.

The next step, I think, is to use qt_configure.bzl to dynamically
determine the paths to these tools on the current system.